+++ uikit/src/main/java/com/sendbird/uikit/internal/model/MessageDisplayDataWrapper.kt
@@ -1,0 +1,8 @@
+package com.sendbird.uikit.internal.model
+
+import com.sendbird.uikit.model.MessageDisplayData
+
+internal data class MessageDisplayDataWrapper(
+    val messageDisplayData: MessageDisplayData,
+    val updatedAt: Long
+)